requirements.txt
streamlit==1.28.1
streamlit-option-menu==0.3.6
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
opencv-python-headless==4.8.1.78
plotly==5.17.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2